types: move MakeVote / MakeBlock functions (#3819)
to the types package

Paritally Fixes #3584
